subscriber/code/requirement.txt
paho-mqtt==1.6.1
web3==5.29.2
eth-tester==0.6.0b6
py-solc-x==1.1.1